+++ zcash_voting/src/prelude.rs
@@ -58,11 +58,11 @@ pub use crate::types::{
     MAX_VOTE_OPTIONS, MIN_PROPOSAL_ID, MIN_VOTE_OPTIONS,
 };
 pub use crate::vote::{
-    commit as commit_vote, parse_recovery, record_submission as record_vote_submission,
-    record_vc_position, recover_commit as recover_vote_commit, recovery_bundle, serialize_recovery,
-    submission as vote_submission, validate_draft_vote, validate_draft_votes, CommittedVote,
-    DraftVote, SignedVoteCommitment, VanWitness, VoteCommit, VoteCommitStage, VoteRecoveryBundle, VoteSigner,
-    VoteSubmission,
+    commit as commit_vote, commit_batch, parse_recovery, record_submission as record_vote_submission,
+    record_vc_position, recover_commit as recover_vote_commit, recover_signed_commitments,
+    recovery_bundle, serialize_recovery, submission as vote_submission, validate_draft_vote,
+    validate_draft_votes, CommittedVote, DraftVote, SignedVoteCommitment, SignedVoteCommitments,
+    VanWitness, VoteCommit, VoteCommitStage, VoteRecoveryBundle, VoteSigner, VoteSubmission,
 };
 pub use crate::warm_proving_caches;